clean up issue with dependency NPEs
