+++ support/cas-server-support-ws-idp/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,2 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.apereo.cas.config.CoreWsSecurityIdentityProviderConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=org.apereo.cas.config.CoreWsSecurityIdentityProviderConfiguration,\
+  org.apereo.cas.config.CoreWsSecurityIdentityProviderWebflowConfiguration
